Move inherent impls around for better docs
