+++ src/lib.rs
@@ -14,16 +14,18 @@
 
 extern crate byteorder;
 
-mod ops;
-mod iter;
 mod store;
 mod container;
 mod util;
-mod serialization;
 mod fmt;
-mod cmp;
 
-use container::Container;
+// Order of these modules matters as it determines the `impl` blocks order in
+// the docs
+mod inherent;
+mod iter;
+mod ops;
+mod cmp;
+mod serialization;
 
 pub use iter::Iter;
 
@@ -45,168 +47,5 @@ pub use iter::Iter;
 /// ```
 #[derive(PartialEq, Clone)]
 pub struct RoaringBitmap {
-    containers: Vec<Container>,
-}
-
-impl RoaringBitmap {
-    /// Creates an empty `RoaringBitmap`.
-    ///
-    /// # Examples
-    ///
-    /// ```rust
-    /// use roaring::RoaringBitmap;
-    /// let mut rb = RoaringBitmap::new();
-    /// ```
-    pub fn new() -> Self {
-        RoaringBitmap { containers: Vec::new() }
-    }
-
-    /// Adds a value to the set. Returns `true` if the value was not already present in the set.
-    ///
-    /// # Examples
-    ///
-    /// ```rust
-    /// use roaring::RoaringBitmap;
-    ///
-    /// let mut rb = RoaringBitmap::new();
-    /// assert_eq!(rb.insert(3), true);
-    /// assert_eq!(rb.insert(3), false);
-    /// assert_eq!(rb.contains(3), true);
-    /// ```
-    pub fn insert(&mut self, value: u32) -> bool {
-        let (key, index) = util::split(value);
-        let container = match self.containers.binary_search_by_key(&key, |c| c.key) {
-            Ok(loc) => &mut self.containers[loc],
-            Err(loc) => {
-                self.containers.insert(loc, Container::new(key));
-                &mut self.containers[loc]
-            },
-        };
-        container.insert(index)
-    }
-
-    /// Removes a value from the set. Returns `true` if the value was present in the set.
-    ///
-    /// # Examples
-    ///
-    /// ```rust
-    /// use roaring::RoaringBitmap;
-    ///
-    /// let mut rb = RoaringBitmap::new();
-    /// rb.insert(3);
-    /// assert_eq!(rb.remove(3), true);
-    /// assert_eq!(rb.remove(3), false);
-    /// assert_eq!(rb.contains(3), false);
-    /// ```
-    pub fn remove(&mut self, value: u32) -> bool {
-        let (key, index) = util::split(value);
-        match self.containers.binary_search_by_key(&key, |c| c.key) {
-            Ok(loc) => {
-                if self.containers[loc].remove(index) {
-                    if self.containers[loc].len == 0 {
-                        self.containers.remove(loc);
-                    }
-                    true
-                } else {
-                    false
-                }
-            }
-            _ => false,
-        }
-    }
-
-    /// Returns `true` if this set contains the specified integer.
-    ///
-    /// # Examples
-    ///
-    /// ```rust
-    /// use roaring::RoaringBitmap;
-    ///
-    /// let mut rb = RoaringBitmap::new();
-    /// rb.insert(1);
-    /// assert_eq!(rb.contains(0), false);
-    /// assert_eq!(rb.contains(1), true);
-    /// assert_eq!(rb.contains(100), false);
-    /// ```
-    pub fn contains(&self, value: u32) -> bool {
-        let (key, index) = util::split(value);
-        match self.containers.binary_search_by_key(&key, |c| c.key) {
-            Ok(loc) => self.containers[loc].contains(index),
-            Err(_) => false,
-        }
-    }
-
-    /// Clears all integers in this set.
-    ///
-    /// # Examples
-    ///
-    /// ```rust
-    /// use roaring::RoaringBitmap;
-    ///
-    /// let mut rb = RoaringBitmap::new();
-    /// rb.insert(1);
-    /// assert_eq!(rb.contains(1), true);
-    /// rb.clear();
-    /// assert_eq!(rb.contains(1), false);
-    /// ```
-    pub fn clear(&mut self) {
-        self.containers.clear();
-    }
-
-    /// Returns `true` if there are no integers in this set.
-    ///
-    /// # Examples
-    ///
-    /// ```rust
-    /// use roaring::RoaringBitmap;
-    ///
-    /// let mut rb = RoaringBitmap::new();
-    /// assert_eq!(rb.is_empty(), true);
-    ///
-    /// rb.insert(3);
-    /// assert_eq!(rb.is_empty(), false);
-    /// ```
-    pub fn is_empty(&self) -> bool {
-        self.containers.is_empty()
-    }
-
-    /// Returns the number of distinct integers added to the set.
-    ///
-    /// # Examples
-    ///
-    /// ```rust
-    /// use roaring::RoaringBitmap;
-    ///
-    /// let mut rb = RoaringBitmap::new();
-    /// assert_eq!(rb.len(), 0);
-    ///
-    /// rb.insert(3);
-    /// assert_eq!(rb.len(), 1);
-    ///
-    /// rb.insert(3);
-    /// rb.insert(4);
-    /// assert_eq!(rb.len(), 2);
-    /// ```
-    pub fn len(&self) -> u64 {
-        self.containers
-            .iter()
-            .map(|container| container.len)
-            .sum()
-    }
-
-    fn min(&self) -> Option<u32> {
-        self.containers.first()
-            .map(|head| util::join(head.key, head.min()))
-    }
-
-    fn max(&self) -> Option<u32> {
-        self.containers.last()
-            .map(|tail| util::join(tail.key, tail.max()))
-    }
-}
-
-impl Default for RoaringBitmap {
-    fn default() -> RoaringBitmap {
-        RoaringBitmap::new()
-    }
+    containers: Vec<container::Container>,
 }
